Move blog directory stuff to src (#86)
